+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/internal/JClassSymbol.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.java.symbols.refs;
+package net.sourceforge.pmd.lang.java.symbols.internal;
 
 import java.lang.reflect.Modifier;
 import java.util.Objects;
