+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/package-info.java
@@ -1,0 +1,16 @@
+/**
+ * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
+ */
+
+/**
+ * Prototype of a new symbol resolution framework
+ * that inter-operates cleanly with type resolution.
+ *
+ * @see net.sourceforge.pmd.lang.java.symbols.internal.JDeclarationSymbol
+ * @see net.sourceforge.pmd.lang.java.symbols.table.JSymbolTable
+ */
+
+@Experimental
+package net.sourceforge.pmd.lang.java.symbols;
+
+import net.sourceforge.pmd.annotation.Experimental;
