+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/table/internal/EmptySymbolTable.java
@@ -7,9 +7,9 @@ package net.sourceforge.pmd.lang.java.symbols.table.internal;
 import java.util.Optional;
 import java.util.stream.Stream;
 
-import net.sourceforge.pmd.lang.java.symbols.refs.JMethodSymbol;
-import net.sourceforge.pmd.lang.java.symbols.refs.JSimpleTypeDeclarationSymbol;
-import net.sourceforge.pmd.lang.java.symbols.refs.JValueSymbol;
+import net.sourceforge.pmd.lang.java.symbols.internal.JMethodSymbol;
+import net.sourceforge.pmd.lang.java.symbols.internal.JSimpleTypeDeclarationSymbol;
+import net.sourceforge.pmd.lang.java.symbols.internal.JValueSymbol;
 import net.sourceforge.pmd.lang.java.symbols.table.JSymbolTable;
