+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/symbols/table/internal/SamePackageSymbolTable.java
@@ -8,9 +8,9 @@ import java.util.Optional;
 import java.util.logging.Logger;
 import java.util.stream.Stream;
 
-import net.sourceforge.pmd.lang.java.symbols.refs.JMethodSymbol;
-import net.sourceforge.pmd.lang.java.symbols.refs.JResolvableClassDeclarationSymbol;
-import net.sourceforge.pmd.lang.java.symbols.refs.JValueSymbol;
+import net.sourceforge.pmd.lang.java.symbols.internal.JMethodSymbol;
+import net.sourceforge.pmd.lang.java.symbols.internal.JResolvableClassDeclarationSymbol;
+import net.sourceforge.pmd.lang.java.symbols.internal.JValueSymbol;
 import net.sourceforge.pmd.lang.java.symbols.table.JSymbolTable;
