+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/symbols/table/HeaderScopesTest.kt
@@ -8,9 +8,9 @@ import javasymbols.testdata.StaticNameCollision
 import javasymbols.testdata.Statics
 import net.sourceforge.pmd.lang.java.ast.parserTest
 import net.sourceforge.pmd.lang.java.symbols.*
-import net.sourceforge.pmd.lang.java.symbols.refs.JFieldSymbol
-import net.sourceforge.pmd.lang.java.symbols.refs.JMethodSymbol
-import net.sourceforge.pmd.lang.java.symbols.refs.JResolvableClassDeclarationSymbol
+import net.sourceforge.pmd.lang.java.symbols.internal.JFieldSymbol
+import net.sourceforge.pmd.lang.java.symbols.internal.JMethodSymbol
+import net.sourceforge.pmd.lang.java.symbols.internal.JResolvableClassDeclarationSymbol
 import net.sourceforge.pmd.lang.java.symbols.table.internal.ImportOnDemandSymbolTable
 import net.sourceforge.pmd.lang.java.symbols.table.internal.JavaLangSymbolTable
 import net.sourceforge.pmd.lang.java.symbols.table.internal.SamePackageSymbolTable
